added partnumbers

+++ src/routes/industries/delete.ts
@@ -3,6 +3,7 @@ import express from "express";
 import { requireAuth } from "../../middlewares";
 import { BadRequestError } from "../../errors";
 import { IndustryRepo } from "../../repos/industry-repo";
+import { ProjectRepo } from "../../repos/project-repo";
 
 const router = express.Router();
 
@@ -14,12 +15,12 @@ router.delete("/industries/:id", requireAuth, async (req, res) => {
     throw new BadRequestError("Industry does not exists");
   }
 
-  // const rfqs = await RfqRepo.findByIndustryId(id);
-  // if (rfqs && rfqs.length > 0) {
-  //   throw new BadRequestError(
-  //     "Industry has RFQs. Delete RFQs first or attach them to another Industry."
-  //   );
-  // }
+  const projects = await ProjectRepo.findByIndustryId(id);
+  if (projects && projects.length > 0) {
+    throw new BadRequestError(
+      "This industry has projects. Delete projects first in order to remove the industry."
+    );
+  }
 
   const deletedIndustry = await IndustryRepo.delete(id);
   res.send(deletedIndustry);
